Let the user confirm the deletion before taking it to action

+++ client/app/js/story_show_dialog.jsx
@@ -52,7 +52,9 @@ export default class StoryShowDialog extends React.Component {
   }
 
   onClickDestroy = () => {
-    BoardActions.deleteStory(this.props.story_id);
+    BoardActions.showConfirmation(() => {
+      BoardActions.deleteStory(this.props.story_id);
+    });
   }
 
   onClickEdit = () => {
